chg: Refine the qrcode component
